add_everything

+++ v2ray-bot/admin_task.py
@@ -2,16 +2,18 @@ import random
 from datetime import datetime, timedelta
 import pytz
 import utilities
-from private import ADMIN_CHAT_ID
-from utilities import ready_report_problem_to_admin, message_to_user, sqlite_manager, api_operation, infinity_name
+from private import ADMIN_CHAT_ID, OTHER_ADMIN
+from utilities import ready_report_problem_to_admin, message_to_user, sqlite_manager, api_operation, infinity_name, report_status_to_admin
 from wallet import WalletManage
 import ranking
+from ranking import rank_access
 
 wallet_manage = WalletManage('User', 'wallet', 'v2ray', 'chat_id')
 ranking_manage = ranking.RankManage('Rank', 'level', 'rank_name',db_name='v2ray', user_id_identifier='chat_id')
 
-
 def admin_add_update_inbound(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     """
     {'update': 3,
     'total_traffic': 0,
@@ -21,8 +23,7 @@ def admin_add_update_inbound(update, context):
     'listen_ip': '',
     'port': 21442,
     'protocol': 'vless'
-    'server_domain': 'admin.ggkala.shop'
-    }
+    'server_domain': 'admin.ggkala.shop'}
     """
     try:
         user_message = eval(update.message.reply_to_message.text)
@@ -57,6 +58,8 @@ def admin_add_update_inbound(update, context):
 
 
 def add_service(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     """
     {
     'update': 0,
@@ -117,6 +120,8 @@ def get_all_service():
 
 
 def all_service(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     try:
         get = get_all_service()[:4000]
         update.message.reply_text('All Service:\n\n' + str(get))
@@ -125,6 +130,8 @@ def all_service(update, context):
 
 
 def del_service(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     if context.args:
         try:
             sqlite_manager.delete({"Product": [context.args[0], eval(context.args[1])]})
@@ -148,7 +155,8 @@ def second_to_ms(date, time_to_ms: bool = True):
         return datetime.fromtimestamp(seconds)
 
 
-def add_client_bot(purchased_id, personalization=None):
+def add_client_bot(purchased_id):
+
     try:
         random_number = random.randint(0, 10_000_000)
         get_client_db = sqlite_manager.select(table='Purchased', where=f'id = {purchased_id}')
@@ -198,6 +206,8 @@ def add_client_bot(purchased_id, personalization=None):
 
 
 def run_in_system(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     try:
         user_message = eval(update.message.reply_to_message.text)
         text = f'ok {user_message}'
@@ -207,6 +217,8 @@ def run_in_system(update, context):
 
 
 def say_to_every_one(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     all_user = sqlite_manager.select('chat_id,name', 'User')
     text = update.message.reply_to_message.text
 
@@ -219,6 +231,8 @@ def say_to_every_one(update, context):
 
 
 def say_to_customer_of_server(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     # user_message = '/say_to_customer_of_server country'
 
     get_server_country = update.message.text.replace('/say_to_customer_of_server ', '')
@@ -243,11 +257,13 @@ def say_to_customer_of_server(update, context):
 
 
 def clear_depleted_service(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     try:
         get_inbound_id = int(update.message.text.replace('/clear_depleted_service ', ''))
 
         customer_service = sqlite_manager.select(column='chat_id,name,client_email,inbound_id', table='Purchased', where=f'status = 0 and inbound_id = {get_inbound_id}')
-        get_server_domain = sqlite_manager.select(column='server_domain', table='Product',
+        sqlite_manager.select(column='server_domain', table='Product',
                                                   where=f'id = "{customer_service[0][3]}"')
 
         reason = update.message.reply_to_message.text if update.message.reply_to_message else 'عدم تمدید و یا ارتقا سرویس'
@@ -265,6 +281,8 @@ def clear_depleted_service(update, context):
 
 
 def add_credit_to_server_customer_wallet(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     # get_admin_order = '/add_credit_to_customer_wallet country, credit'
     try:
         get_admin_order = update.message.text.replace('/add_credit_to_customer_wallet ', '').split(', ')
@@ -297,6 +315,8 @@ def add_credit_to_server_customer_wallet(update, context):
 
 
 def add_credit_to_customer(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
     try:
         get_admin_order = update.message.text.replace('/add_credit_to_customer ', '').split(', ')
         get_user_chat_id = get_admin_order[0]
@@ -320,6 +340,7 @@ def add_credit_to_customer(update, context):
 
 
 def check_all_configs(chat_id, inbound_id, product_id=None):
+
     if not product_id:
         product_id = sqlite_manager.select(column='id', table='Product', limit=1)[0][0]
     get_all = api_operation.get_all_inbounds()
@@ -335,21 +356,32 @@ def check_all_configs(chat_id, inbound_id, product_id=None):
 # check_all_configs(6450325872, 5)
 
 def admin_rank_up(update, context):
+    chat_id = update.message.chat_id
+    if chat_id not in OTHER_ADMIN: return
+
+    get_admin_order = update.message.text.replace('/rank_up ', '').split(', ')
+    get_user_chat_id = get_admin_order[0]
     try:
-        get_admin_order = update.message.text.replace('/rank_up ', '').split(', ')
-        get_user_chat_id = get_admin_order[0]
         get_rank_name = get_admin_order[1]
-        rank_access = '\n'.join(ranking.rank_access_fa[get_rank_name][1:])
+        rank_access_ = '\n'.join(ranking.rank_access_fa[get_rank_name][1:])
 
-        text = f'رنک شما توسط ادمین ارتقا یافت.\n\nویژگی های این رنک:\n {rank_access}'
+        text = f'رنک شما توسط ادمین ارتقا یافت.\n\nویژگی های این رنک:\n {rank_access_}'
 
-        customer_of_service = sqlite_manager.select(column='name,user_name', table='User',
+        sqlite_manager.select(column='name,user_name', table='User',
                                                     where=f'chat_id = {get_user_chat_id}')
 
         ranking_manage.rank_up(get_rank_name, get_user_chat_id)
         message_to_user(update, context, message=text, chat_id=get_user_chat_id)
 
         update.message.reply_text('RANKUP SUCCESS')
+
+    except TypeError:
+        rank_name_ = next(iter(rank_access))
+        level = 0
+
+        sqlite_manager.insert(table='Rank', rows=[{'name': None, 'user_name': None, 'chat_id': get_user_chat_id,
+                                                   'level': level, 'rank_name': rank_name_}])
+        report_status_to_admin(context, 'I Create Rank For User. Try Again!', get_user_chat_id)
 
     except Exception as e:
         ready_report_problem_to_admin(context, 'admin_rank_up', update.message.from_user['id'], e)
